Revert "Move non edgeql related directives/roles into new `edb` domain"
