+++ docs/clients/graphql/index.rst
@@ -1,4 +1,4 @@
-.. edb:section-intro-page:: graphql
+.. eql:section-intro-page:: graphql
 
 .. _ref_graphql_index:
