+++ docs/clients/index.rst
@@ -1,4 +1,4 @@
-.. edb:section-intro-page:: clients
+.. eql:section-intro-page:: clients
 
 .. _ref_clients_index:
