+++ docs/edgeql/index.rst
@@ -1,4 +1,4 @@
-.. edb:section-intro-page:: edgeql
+.. eql:section-intro-page:: edgeql
 .. versioned-section::
 
 .. _ref_edgeql:
