+++ docs/reference/index.rst
@@ -1,4 +1,4 @@
-.. edb:section-intro-page:: reference
+.. eql:section-intro-page:: reference
 .. versioned-section::
 
 .. _ref_reference_index:
