+++ docs/stdlib/index.rst
@@ -1,4 +1,4 @@
-.. edb:section-intro-page:: stdlib
+.. eql:section-intro-page:: stdlib
 .. versioned-section::
 
 .. _ref_std:
